Get sprites and ride overall view working in large maps

+++ src/openrct2/world/SpriteBase.h
@@ -30,18 +30,18 @@ struct SpriteBase
     uint8_t sprite_height_negative;
     uint16_t sprite_index;
     uint16_t flags;
-    int16_t x;
-    int16_t y;
-    int16_t z;
+    int32_t x;
+    int32_t y;
+    int32_t z;
     // Width from centre of sprite to edge
     uint8_t sprite_width;
     // Height from centre of sprite to top
     uint8_t sprite_height_positive;
     // Screen Coordinates of sprite
-    int16_t sprite_left;
-    int16_t sprite_top;
-    int16_t sprite_right;
-    int16_t sprite_bottom;
+    int32_t sprite_left;
+    int32_t sprite_top;
+    int32_t sprite_right;
+    int32_t sprite_bottom;
 
     uint8_t sprite_direction;
